third_party/WebKit/LayoutTests/fast/files/blob-close-read-expected.txt

Test the Blob.close() method, reading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing that the reading of closed Blobs fail.
PASS blob.close(); blob.size is 0
PASS reader.readAsArrayBuffer(blob) threw exception InvalidStateError: Failed to execute 'readAsArrayBuffer' on 'FileReader': Blob has been closed..
PASS reader.readAsBinaryString(blob) threw exception InvalidStateError: Failed to execute 'readAsBinaryString' on 'FileReader': Blob has been closed..
PASS reader.readAsText(blob) threw exception InvalidStateError: Failed to execute 'readAsText' on 'FileReader': Blob has been closed..
PASS reader.readAsDataURL(blob) threw exception InvalidStateError: Failed to execute 'readAsDataURL' on 'FileReader': Blob has been closed..
Testing that ongoing async reads aren't interrupted by close()
PASS reader2.readAsArrayBuffer(blob) threw exception InvalidStateError: Failed to execute 'readAsArrayBuffer' on 'FileReader': Blob has been closed..
PASS blob.size is 0
PASS FileReader loaded: 0123456789abcdef
PASS readAsText() completed
PASS blob.size is 0
Testing that sliced reads aren't affected by close() on 'parent' Blob.
PASS sliced.size is 14
PASS result is "23456789abcdef"
PASS readAsText() completed
PASS sliced.size is 14
PASS successfullyParsed is true

TEST COMPLETE
